refactor: removed all implicit conversions of into -> float

+++ headers/sprite.hpp
@@ -6,16 +6,16 @@
 struct Sprite
 {
     Texture2D Texture{};
-    float MaxFramesX{};
-    float MaxFramesY{};
-    float FrameX{};
-    float FrameY{};
+    int MaxFramesX{};
+    int MaxFramesY{};
+    int FrameX{};
+    int FrameY{};
     int PreviousFrame{};
     float UpdateTime{1.f/8.f};
     float RunningTime{};
 
     Sprite() = default;
-    explicit Sprite(const Texture2D& Texture, float MaxFramesX = 1, float MaxFramesY = 1, float UpdateSpeed = 1.f/8.f);
+    explicit Sprite(const Texture2D& Texture, float MaxFramesX = 1.f, float MaxFramesY = 1.f, float UpdateSpeed = 1.f/8.f);
     void Tick(float DeltaTime);
     Rectangle GetSourceRec();
     Rectangle GetPosRec(const Vector2& ScreenPos, float Scale);
